\name{insert}
\alias{insert}

\title{
Inserting a value into a vector
}

\description{
This inserts a value \code{x0} at \code{d}-th position of \code{x}.
}

\usage{
insert(x, d, x0 = 0)
}

\arguments{
  \item{x}{A vector.}
  \item{d}{A position in \code{x}.}
  \item{x0}{A value to be inserted.}
}

\value{
  The resulting vector.
}


\keyword{free energy}
\keyword{expectation}


